#108 Energy increases by 50% on switching to Night cycle

+++ lib/widgets/energy_bar.dart
@@ -7,16 +7,13 @@ class EnergyBar extends StatefulWidget{
 }
 
 class _EnergyBarState extends State<EnergyBar>{
-  int _energy;
-
+  double _energy;
 
   @override
   void initState(){
     super.initState();
     _energy = Energy.energyInit;
   }
-
-
 
   Widget _makeText() {
     _energy = Energy.energy;
